# api/docs/swagger.yaml
definitions:
...
  editory_submission_genproto_content_service.CreateJournalReq:
    properties:
      acceptance_rate:
        type: string
      acceptance_to_publication:
        type: string
      access:
        type: boolean
      author:
        type: string
      citation_indicator:
        type: string
      cover_photo:
        type: string
      description:
        type: string
      impact_factor:
        type: string
      isbn:
        type: string
      journal_data:
        items:
          $ref: '#/definitions/editory_submission_genproto_content_service.JournalData'
        type: array
      price:
        type: integer
      status:
        type: string
      subjects:
        items:
          $ref: '#/definitions/editory_submission_genproto_content_service.Subject'
        type: array
      submission_to_final_decision:
        type: string
      title:
        type: string
    type: object
...
  editory_submission_genproto_content_service.GetJournalListRes:
    properties:
      count:
        type: integer
      journals:
        items:
          $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
        type: array
    type: object
...
  editory_submission_genproto_content_service.Journal:
    properties:
      acceptance_rate:
        type: string
      acceptance_to_publication:
        type: string
      access:
        type: boolean
      author:
        type: string
      citation_indicator:
        type: string
      cover_photo:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      impact_factor:
        type: string
      isbn:
        type: string
      journal_data:
        items:
          $ref: '#/definitions/editory_submission_genproto_content_service.JournalData'
        type: array
      price:
        type: integer
      status:
        type: string
      subjects:
        items:
          $ref: '#/definitions/editory_submission_genproto_content_service.Subject'
        type: array
      submission_to_final_decision:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  editory_submission_genproto_content_service.JournalData:
    properties:
      journal_id:
        type: string
      short_text:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
  editory_submission_genproto_content_service.Subject:
    properties:
      id:
        type: string
      title:
        type: string
    type: object
...
  http.Header:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
  http.Response:
    properties:
      body:
        description: |-
          Body represents the response body.

          The response body is streamed on demand as the Body field
          is read. If the network connection fails or the server
          terminates the response, Body.Read calls return an error.

          The http Client and Transport guarantee that Body is always
          non-nil, even on responses without a body or responses with
          a zero-length body. It is the caller's responsibility to
          close Body. The default HTTP client's Transport may not
          reuse HTTP/1.x "keep-alive" TCP connections if the Body is
          not read to completion and closed.

          The Body is automatically dechunked if the server replied
          with a "chunked" Transfer-Encoding.

          As of Go 1.12, the Body will also implement io.Writer
          on a successful "101 Switching Protocols" response,
          as used by WebSockets and HTTP/2's "h2c" mode.
      close:
        description: |-
          Close records whether the header directed that the connection be
          closed after reading Body. The value is advice for clients: neither
          ReadResponse nor Response.Write ever closes a connection.
        type: boolean
      contentLength:
        description: |-
          ContentLength records the length of the associated content. The
          value -1 indicates that the length is unknown. Unless Request.Method
          is "HEAD", values >= 0 indicate that the given number of bytes may
          be read from Body.
        type: integer
      header:
        allOf:
        - $ref: '#/definitions/http.Header'
        description: |-
          Header maps header keys to values. If the response had multiple
          headers with the same key, they may be concatenated, with comma
          delimiters.  (RFC 7230, section 3.2.2 requires that multiple headers
          be semantically equivalent to a comma-delimited sequence.) When
          Header values are duplicated by other fields in this struct (e.g.,
          ContentLength, TransferEncoding, Trailer), the field values are
          authoritative.

          Keys in the map are canonicalized (see CanonicalHeaderKey).
      proto:
        description: e.g. "HTTP/1.0"
        type: string
      protoMajor:
        description: e.g. 1
        type: integer
      protoMinor:
        description: e.g. 0
        type: integer
      request:
        allOf:
        - $ref: '#/definitions/http.Request'
        description: |-
          Request is the request that was sent to obtain this Response.
          Request's Body is nil (having already been consumed).
          This is only populated for Client requests.
      status:
        description: e.g. "200 OK"
        type: string
      statusCode:
        description: e.g. 200
        type: integer
      tls:
        allOf:
        - $ref: '#/definitions/tls.ConnectionState'
        description: |-
          TLS contains information about the TLS connection on which the
          response was received. It is nil for unencrypted responses.
          The pointer is shared between responses and should not be
          modified.
      trailer:
        allOf:
        - $ref: '#/definitions/http.Header'
        description: |-
          Trailer maps trailer keys to values in the same
          format as Header.

          The Trailer initially contains only nil values, one for
          each key specified in the server's "Trailer" header
          value. Those values are not added to Header.

          Trailer must not be accessed concurrently with Read calls
          on the Body.

          After Body.Read has returned io.EOF, Trailer will contain
          any trailer values sent by the server.
      transferEncoding:
        description: |-
          Contains transfer encodings from outer-most to inner-most. Value is
          nil, means that "identity" encoding is used.
        items:
          type: string
        type: array
      uncompressed:
        description: |-
          Uncompressed reports whether the response was sent compressed but
          was decompressed by the http package. When true, reading from
          Body yields the uncompressed content instead of the compressed
          content actually set from the server, ContentLength is set to -1,
          and the "Content-Length" and "Content-Encoding" fields are deleted
          from the responseHeader. To get the original response from
          the server, set Transport.DisableCompression to true.
        type: boolean
    type: object
...
  tls.ConnectionState:
    properties:
      cipherSuite:
        description: |-
          CipherSuite is the cipher suite negotiated for the connection (e.g.
          TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256, TLS_AES_128_GCM_SHA256).
        type: integer
      didResume:
        description: |-
          DidResume is true if this connection was successfully resumed from a
          previous session with a session ticket or similar mechanism.
        type: boolean
      handshakeComplete:
        description: HandshakeComplete is true if the handshake has concluded.
        type: boolean
      negotiatedProtocol:
        description: NegotiatedProtocol is the application protocol negotiated with
          ALPN.
        type: string
      negotiatedProtocolIsMutual:
        description: |-
          NegotiatedProtocolIsMutual used to indicate a mutual NPN negotiation.

          Deprecated: this value is always true.
        type: boolean
      ocspresponse:
        description: |-
          OCSPResponse is a stapled Online Certificate Status Protocol (OCSP)
          response provided by the peer for the leaf certificate, if any.
        items:
          type: integer
        type: array
      peerCertificates:
        description: |-
          PeerCertificates are the parsed certificates sent by the peer, in the
          order in which they were sent. The first element is the leaf certificate
          that the connection is verified against.

          On the client side, it can't be empty. On the server side, it can be
          empty if Config.ClientAuth is not RequireAnyClientCert or
          RequireAndVerifyClientCert.
        items:
          $ref: '#/definitions/x509.Certificate'
        type: array
      serverName:
        description: |-
          ServerName is the value of the Server Name Indication extension sent by
          the client. It's available both on the server and on the client side.
        type: string
      signedCertificateTimestamps:
        description: |-
          SignedCertificateTimestamps is a list of SCTs provided by the peer
          through the TLS handshake for the leaf certificate, if any.
        items:
          items:
            type: integer
          type: array
        type: array
      tlsunique:
        description: |-
          TLSUnique contains the "tls-unique" channel binding value (see RFC 5929,
          Section 3). This value will be nil for TLS 1.3 connections and for all
          resumed connections.

          Deprecated: there are conditions in which this value might not be unique
          to a connection. See the Security Considerations sections of RFC 5705 and
          RFC 7627, and https://mitls.org/pages/attacks/3SHAKE#channelbindings.
        items:
          type: integer
        type: array
      verifiedChains:
        description: |-
          VerifiedChains is a list of one or more chains where the first element is
          PeerCertificates[0] and the last element is from Config.RootCAs (on the
          client side) or Config.ClientCAs (on the server side).

          On the client side, it's set if Config.InsecureSkipVerify is false. On
          the server side, it's set if Config.ClientAuth is VerifyClientCertIfGiven
          (and the peer provided a certificate) or RequireAndVerifyClientCert.
        items:
          items:
            $ref: '#/definitions/x509.Certificate'
          type: array
        type: array
      version:
        description: Version is the TLS version used by the connection (e.g. VersionTLS12).
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
...
  /admin/journal:
    get:
      consumes:
      - application/json
      description: Get Journal List
      operationId: get_admin_journal_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetJournalListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.GetJournalListRes'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Journal List
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: 'Create Journal types: [''EDITOR_SPOTLIGHT'', ''SPECIAL_ISSUE'',
        ''ABOUT_JOURNAL'', ''EDITORIAL_BARD'', ''PEER_REVIEW_PROCESS'', ''PUBLICATION_ETHICS'',
        ''ABSTRACTING_INDEXING'', ''ARTICLE_PROCESSING_CHARGES'']'
      operationId: create_admin_journal
      parameters:
      - description: CreateJournalRequestBody
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.CreateJournalReq'
      produces:
      - application/json
      responses:
        "201":
          description: Journal data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Journal
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Update Journal
      operationId: update_admin_journal
      parameters:
      - description: UpdateJournalRequestBody
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
      produces:
      - application/json
      responses:
        "200":
          description: Journal data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Journal
      tags:
      - Admin
  /admin/journal/{journal-id}:
    delete:
      consumes:
      - application/json
      description: Get Journal
      operationId: delete_admin_journal
      parameters:
      - description: journal-id
        in: path
        name: journal-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Journal
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get Journal By ID
      operationId: get_admin_journal_by_id
      parameters:
      - description: journal-id
        in: path
        name: journal-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JournalBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Journal By ID
      tags:
      - Admin
...
  /journal:
    put:
      consumes:
      - application/json
      description: Update Journal
      operationId: update_journal
      parameters:
      - description: UpdateJournalRequestBody
        in: body
        name: journal
        required: true
        schema:
          $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
      produces:
      - application/json
      responses:
        "200":
          description: Journal data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Journal
      tags:
      - Journal
  /journal/{journal-id}:
    delete:
      consumes:
      - application/json
      description: Get Journal
      operationId: delete_journal
      parameters:
      - description: journal-id
        in: path
        name: journal-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Journal
      tags:
      - Journal
    get:
      consumes:
      - application/json
      description: Get Journal By ID
      operationId: get_journal_by_id
      parameters:
      - description: journal-id
        in: path
        name: journal-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: JournalBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/editory_submission_genproto_content_service.Journal'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Journal By ID
      tags:
      - Journal
...
swagger: "2.0"